provide working getcwd() implementation
If you do an _ecos_getcwd(), you get a path like "/hda2/GAME"; if you try
to chdir to that or open a file with that prefix, it won't work, because
the real path is "/fat20a2/hda2/GAME".
eCos has several methods of tracking the current working directory, and
it seems that different firmwares are configured to use different methods,
so this problem may not manifest itself everywhere.