libgame: detect available memory
Not all memory from 0xa00000 to the end of RAM is reserved for the user;
the upper regions contain framebuffers and in some firmwares a NAND cache,
and possibly other stuff owned by the operating system. We find the
location where the maximum size of the user program is kept and use it to
adjust the top-of-heap used by sbrk().